Add files via upload
I am going to be portraying machine learning strategies utilized on an expansive dataset to analyze lodging advertisements in multiple cities in Pakistan. Pakistan could be an exceptionally crowded nation, as of now having the fifth-highest populace out of all nations. The nation has more than 220 million individuals with an annual development rate of 2%. This annual altar is on the higher conclusion. For comparison, the rates of the three most crowded nations, China, India, and the United States are 0.39%, 0.99%, and 0.59% separately. It is evident to see that the lodging advertisement in Pakistan is growing, and will proceed to extend. Understanding the patterns in this showcase, and which components cause those patterns, is basic for the Pakistani Government, arrive designers, genuine bequest organizations, and genuine domain investors. The objective of my examination is to compare the execution of distinctive machine learning models, all anticipating the costs of houses based on known components. I'll construct distinctive sorts of models and see how the results compare. I will also look at how different factors impact the price of a house, and which factors, in particular, are the most impactful, and the least impactful. The following machine learning methods will be tested throughout this paper: Linear Regression, Decision Trees, and Random Forest. I will split our data into disjoint training and testing sets in order to have accurate assessments and reliable results. This will allow us to fairly judge the performance of a specific model. Whenever possible, I will try to employ techniques to improve the quality of our model.